Support fractional pose/image stamps and richer object detections.
Location/ImageMessage timestamps are float; objects store robot_id and capture time, ignore LLM detections on the map path, and prune Ignite entries older than 15s.

+++ dds/dds_utils/cli_log.py
@@ -51,10 +51,10 @@ class ImageLogThrottle:
         self.summary_interval_sec = summary_interval_sec
         self._window_count = 0
         self._window_start = time.time()
-        self._last_ts: int | None = None
+        self._last_ts: float | None = None
         self._logged_first = False
 
-    def record(self, component: str, timestamp: int) -> None:
+    def record(self, component: str, timestamp: float) -> None:
         self._last_ts = timestamp
         now = time.time()
         if is_dds_verbose():
